+++ lib/active_record/connection_adapters/trino_adapter.rb
@@ -32,6 +32,8 @@ module ActiveRecord
         @client_options = ActiveRecord::Trino::Config.client_options(@config)
         @slow_query_threshold = ActiveRecord::Trino::Config.slow_query_threshold(@config)
         @persistent = ActiveRecord::Trino::Config.persistent?(@config)
+        @bulk_column_reflection = ActiveRecord::Trino::Config.bulk_column_reflection?(@config)
+        @static_schema = ActiveRecord::Trino::Config.static_schema?(@config)
         @client = build_client
         install_safety_belts!
       end
@@ -53,6 +55,7 @@ module ActiveRecord
         @persistent_faraday&.close
         @persistent_faraday = nil
         @client = nil
+        clear_column_cache!
       end
 
       def supports_transactions?
@@ -112,6 +115,14 @@ module ActiveRecord
 
       def persistent?
         @persistent
+      end
+
+      def bulk_column_reflection?
+        @bulk_column_reflection
+      end
+
+      def static_schema?
+        @static_schema
       end
 
     private