+++ lib/active_record/trino/config.rb
@@ -37,6 +37,14 @@ module ActiveRecord
         !!symbolize(config).fetch(:persistent, false)
       end
 
+      def bulk_column_reflection?(config)
+        !!symbolize(config).fetch(:bulk_column_reflection, false)
+      end
+
+      def static_schema?(config)
+        !!symbolize(config).fetch(:static_schema, false)
+      end
+
       def default_port(ssl)
         ssl ? DEFAULT_HTTPS_PORT : DEFAULT_HTTP_PORT
       end